rename third_party
